Build updates (#19)
* Update workflow to user repo

* Upgrade to UBI 9 and use go install

Upgrade to UBI 9 to have Python 3.9 as the primary and only Python
version in the container. This removes the dependency on a vulnerable
version of pip, version 9.0.3, pulled in from installing git with
Python 3.6.8. The switch to the go-toolset builder image required no
longer building in the root directory as well as fully deprecated go
get for installation. go install is used specifying the GOBIN to be
the temporary $SEC_BIN defined for the builder.

Addition and usage of a quick-instALL.sh script built from instALL.sh
which combines all the microdnf install commands into a single
command. This removes the slow repetitive query done at the start of
each install command.

Combined RUN commands in the Dockerfile to reduce the number of layers
in the built image, allowing faster pulls of the image.

Signed-off-by: Patrick Hickey <pahickey@redhat.com>

* Fix Workflow